requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
Authlib==1.3.2
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distro==1.9.0
Flask==3.0.3
Flask-Session==0.8.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
msgspec==0.18.6
multidict==6.1.0
numpy==2.1.1
openai==0.28.0
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyVirtualDisplay==3.0
redis==5.1.1
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
xlib==0.21
yarl==1.13.1
